Ptzp: MgeoGungorHead: Bug fixed.
This camera used different tcp channel for each head. So we need to return

meslen for uncorrect read bytes. This solution fixed multi problem for thermal cam.

The other problem is sync not completed because of our timer not restart. It's fixed.

We were sending data very often. This situation blocked the dataReady. After fixed

we are sending data in 100 ms.